fix: split agy provider tests to satisfy backend-size limit
Move ScriptWrapProvider Agy tests into inference_provider_agy_test.go so inference_provider_streaming_test.go stays under the 1000-line cap.
